+++ src/Sockets.php
@@ -3,13 +3,11 @@ declare(strict_types = 1);
 
 namespace Innmind\OperatingSystem;
 
-use Innmind\IO\Sockets\{
-    Client,
-    Server,
+use Innmind\IO\{
+    Sockets\Clients\Client,
+    Sockets\Servers\Server,
+    Sockets\Unix\Address,
 };
-use Innmind\Socket\Address\Unix;
-use Innmind\TimeContinuum\ElapsedPeriod;
-use Innmind\Stream\Watch;
 use Innmind\Immutable\Maybe;
 
 interface Sockets
@@ -19,18 +17,17 @@ interface Sockets
      *
      * @return Maybe<Server>
      */
-    public function open(Unix $address): Maybe;
+    public function open(Address $address): Maybe;
 
     /**
      * This will take control of the socket if it already exist (use carefully)
      *
      * @return Maybe<Server>
      */
-    public function takeOver(Unix $address): Maybe;
+    public function takeOver(Address $address): Maybe;
 
     /**
      * @return Maybe<Client>
      */
-    public function connectTo(Unix $address): Maybe;
-    public function watch(?ElapsedPeriod $timeout = null): Watch;
+    public function connectTo(Address $address): Maybe;
 }